Move EventFilter/Utilities/plugins/testGlobalNumbers.cc to EventFilter/Utilities/test/testGlobalNumbers.cc, and add a BuildFile in EventFilter/Utilities/test.
