+++ EventFilter/Utilities/test/BuildFile.xml
@@ -1,0 +1,10 @@
+<library   file="*.cc" name="EventFilterUtilitiesTest">
+  <use   name="FWCore/Framework"/>
+  <use   name="FWCore/MessageLogger"/>
+  <use   name="FWCore/ParameterSet"/>
+  <use   name="FWCore/Utilities"/>
+  <use   name="DataFormats/FEDRawData"/>
+  <use   name="EventFilter/FEDInterface"/>
+  <use   name="EventFilter/Utilities"/>
+  <flags   EDM_PLUGIN="1"/>
+</library>
